Fix even more PMD error
